Add files via upload
This is the new version, we struggle still with margins and also the columns are meant to be side by side. However we came a long way from the previous version. we added a footer, we got food and prices with information. Opening hours etc. Hope this is enough to pass, we shall continue to perfect our skills for exam season. 

Group: 97
s376867 Mert Altunbas
s353670 Hashim Ul Hassan
S374918 - Alex McCorkle
S321812 - Andreas Skuggen